Moved posts.py to posts_view.py
