+++ json-server-py
@@ -4,8 +4,7 @@ from nss_handler import HandleRequests, status
 
 
 # Add your imports below this line
-from views import create_user, login_user, getAllPosts, get_all_tags, create_tag
-from views import retrieve_myposts, getSinglePost
+from views import create_user, login_user, getAllPosts, get_all_tags, create_tag, retrieve_myposts, getSinglePost, create_post
 
 
 class JSONServer(HandleRequests):
@@ -33,13 +32,11 @@ class JSONServer(HandleRequests):
             if url["pk"] != 0:
                 response_body = retrieve_myposts(pk, url)
                 return self.response(response_body, status.HTTP_200_SUCCESS.value)
-
-
-
         if url["requested_resource"] == "myposts":
             if url["pk"] != 0:
                 response_body = retrieve_myposts(pk, url)
                 return self.response(response_body, status.HTTP_200_SUCCESS.value)
+
 
 
 
@@ -53,7 +50,7 @@ class JSONServer(HandleRequests):
         """Handle POST requests from a client"""
 
         url = self.parse_url(self.path)
-        pk = url["pk"]
+        
 
         content_len = int(self.headers.get("content-length", 0))
         request_body = self.rfile.read(content_len)
@@ -78,7 +75,15 @@ class JSONServer(HandleRequests):
                     status.HTTP_404_CLIENT_ERROR_RESOURCE_NOT_FOUND.value,
                 )
             
-        elif url["requested_resource"] == "tags":
+        if url["requested_resource"] == "posts":
+            new_post_id = create_post(request_body)
+            if new_post_id:
+                return self.response(
+                    json.dumps({ "message": "Post created", "post_id": new_post_id }),
+                    status.HTTP_201_SUCCESS_CREATED.value
+                )
+                  
+        if url["requested_resource"] == "tags":
             created = create_tag(request_body)
             if created:
                 return self.response(created, status.HTTP_201_SUCCESS_CREATED.value)
@@ -86,7 +91,8 @@ class JSONServer(HandleRequests):
                 return self.response(
                     json.dumps({ "error": "Failed to create tag" }),
                     status.HTTP_404_CLIENT_ERROR_RESOURCE_NOT_FOUND.value,
-                )  
+                )
+            
 
 
 def main():
