+++ views/__init__.py
@@ -1,4 +1,3 @@
-from .posts_view import getAllPosts, getSinglePost, retrieve_myposts
+from .posts_view import getAllPosts, getSinglePost, retrieve_myposts, create_post
 from .tag_views import get_all_tags, create_tag
 from .user import create_user, login_user
-from .post import create_post
